Increase browser launch timeout in all tests + bundle tests config refactor
Put all the configuration in the configuration helper to be able to share the
same configuration for all kinds of test.
